Login component as per new structure.
