+++ packages/authentication/src/index.js
@@ -11,3 +11,10 @@ export { AuthenticationMiddlewareGenerator } from './AuthenticationMiddleware/Au
 export { DEFAULT_CONFIG_AUTH_MIDDLEWARE } from './AuthenticationMiddleware/config';
 export { accessLevelTypes, userRoles, membershipStatus } from './utils/enums';
 export { default as redirect } from './utils/redirect';
+
+// Login exports
+export { LoginGenerator } from './Login/LoginGenerator';
+export { DEFAULT_CONFIG_LOGIN } from './Login/config';
+export { default as useQuery } from './Login/hooks/useQuery';
+export { default as getIdps } from './Login/utils/getIdps';
+export { default as getRedirectPath } from './Login/utils/getRedirectPath';
