--- packages/bento-frontend/src/pages/accessManagment/login/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import LoginView from './view';
-
-function loginController(props) {
-  return (
-    <div>
-      <LoginView {...props} />
-    </div>
-  );
-}
-
-export default loginController;
